Revert Change in TrMath

+++ TR5Main/Global/trmath.cpp
@@ -1311,17 +1311,17 @@ unsigned short atanTab[] = {
 // fix "expression must have integral or unscoped enum type"
 short ANGLE(float angle)
 {
-	return short(angle * 65536.0f / 360.0f);
+	return angle * 65536.0f / 360.0f;
 }
 
 float TO_DEGREES(short angle)
 {
-	return float(angle) * 360.0f / 65536.0f;
+	return (unsigned short) angle * 360.0f / 65536.0f;
 }
 
 float TO_RAD(short angle)
 {
-	return float(angle) * 360.0f / 65536.0f * RADIAN;
+	return angle * 360.0f / 65536.0f * RADIAN;
 }
 
 const float frand()
